+++ src/main/java/fll/web/admin/SubjectiveScheduleByCategory.java
@@ -18,11 +18,9 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import javax.sql.DataSource;
 
-import com.itextpdf.text.DocumentException;
-
 import fll.db.Queries;
+import fll.scheduler.ScheduleWriter;
 import fll.scheduler.TournamentSchedule;
-import fll.util.FLLInternalException;
 import fll.web.ApplicationAttributes;
 import fll.web.BaseFLLServlet;
 import fll.web.SessionAttributes;
@@ -61,11 +59,7 @@ public class SubjectiveScheduleByCategory extends BaseFLLServlet {
       response.reset();
       response.setContentType("application/pdf");
       response.setHeader("Content-Disposition", "filename=subjectiveByCategoryAndTime.pdf");
-      schedule.outputSubjectiveSchedulesByCategory(response.getOutputStream());
-
-    } catch (final DocumentException e) {
-      LOGGER.error(e.getMessage(), e);
-      throw new FLLInternalException("Got error writing schedule", e);
+      ScheduleWriter.outputSubjectiveSchedulesByCategory(schedule, response.getOutputStream());
     } catch (final SQLException sqle) {
       LOGGER.error(sqle.getMessage(), sqle);
       throw new RuntimeException("Error saving team data into the database", sqle);