+++ src/main/java/fll/web/report/FinalComputedScores.java
@@ -220,8 +220,10 @@ public final class FinalComputedScores extends BaseFLLServlet {
     rootElement.appendChild(layoutMasterSet);
 
     final String pageMasterName = "simple";
-    FOPUtils.createSimplePageMaster(document, layoutMasterSet, pageMasterName, FOPUtils.PAGE_LETTER_SIZE,
-                                    FOPUtils.STANDARD_MARGINS, HEADER_MARGIN_INCHES, FOOTER_MARGIN_INCHES);
+    final Element pageMaster = FOPUtils.createSimplePageMaster(document, pageMasterName, FOPUtils.PAGE_LETTER_SIZE,
+                                                               FOPUtils.STANDARD_MARGINS, HEADER_MARGIN_INCHES,
+                                                               FOOTER_MARGIN_INCHES);
+    layoutMasterSet.appendChild(pageMaster);
 
     final Element legend = createLegend(document, percentageHurdle, standardMean, standardSigma);
 
@@ -560,7 +562,7 @@ public final class FinalComputedScores extends BaseFLLServlet {
             overallScore = ts;
           }
 
-          final Element row1 = FOPUtils.createXslFoElement(document, "table-row");
+          final Element row1 = FOPUtils.createTableRow(document);
           tableBody.appendChild(row1);
           row1.setAttribute("keep-with-next", "always");
 
@@ -581,7 +583,7 @@ public final class FinalComputedScores extends BaseFLLServlet {
           row1.appendChild(FOPUtils.createTableCell(document, null, ""));
 
           // The second row of the team table...
-          final Element row2 = FOPUtils.createXslFoElement(document, "table-row");
+          final Element row2 = FOPUtils.createTableRow(document);
           tableBody.appendChild(row2);
 
           // First column contains the team # and name
@@ -763,7 +765,7 @@ public final class FinalComputedScores extends BaseFLLServlet {
 
     final Element tableHeader = FOPUtils.createXslFoElement(document, "table-header");
 
-    final Element row1 = FOPUtils.createXslFoElement(document, "table-row");
+    final Element row1 = FOPUtils.createTableRow(document);
     tableHeader.appendChild(row1);
     row1.setAttribute("font-weight", "bold");
     row1.setAttribute("keep-with-next", "always");
@@ -789,7 +791,7 @@ public final class FinalComputedScores extends BaseFLLServlet {
 
     // row 2 needs a bottom border, the border is added to each cell
 
-    final Element row2 = FOPUtils.createXslFoElement(document, "table-row");
+    final Element row2 = FOPUtils.createTableRow(document);
     tableHeader.appendChild(row2);
 
     final Element teamNumCell = FOPUtils.createTableCell(document, null, "Team # / Team Name");
@@ -964,7 +966,7 @@ public final class FinalComputedScores extends BaseFLLServlet {
     final Element tableBody = FOPUtils.createXslFoElement(document, "table-body");
     header.appendChild(tableBody);
 
-    final Element row1 = FOPUtils.createXslFoElement(document, "table-row");
+    final Element row1 = FOPUtils.createTableRow(document);
     tableBody.appendChild(row1);
     row1.setAttribute("keep-with-next", "always");
 
@@ -982,7 +984,7 @@ public final class FinalComputedScores extends BaseFLLServlet {
     row1.appendChild(titleCell);
     titleCell.setAttribute("padding-right", sidePadding);
 
-    final Element row2 = FOPUtils.createXslFoElement(document, "table-row");
+    final Element row2 = FOPUtils.createTableRow(document);
     tableBody.appendChild(row2);
 
     final Element finalScoresCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_LEFT,